requirements.txt
amqp
asgiref
billiard
click
click-didyoumean
click-plugins
click-repl
colorama
crispy-bootstrap5
Django~=4.2.2
django-crispy-forms~=2.0
kombu
Pillow
prompt-toolkit
psycopg2-binary
pytz
six
sqlparse
tzdata
Unidecode
vine
wcwidth
redis
python-dotenv~=1.0.0
dotenv~=0.0.5
schedule~=1.2.0